requirements.txt
annotated-types==0.7.0
certifi==2025.10.5
charset-normalizer==3.4.4
gds==0.0.2
graphdatascience==1.17
idna==3.11
kafka-python==2.2.15
multimethod==2.0
neo4j==5.28.2
numpy==2.2.6
pandas==2.3.3
pyarrow==20.0.0
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
sseclient-py==1.8.0
tenacity==9.1.2
textdistance==4.6.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
